requirements.txt
aiogram==3.26.0
aiohttp==3.13.3
datasets==2.20.0
mlx_lm==0.21.0
openai==2.29.0
pandas==3.0.1
psutil==7.2.1
python-dotenv==1.2.2
PyYAML==6.0.3
sentence_transformers==5.3.0
streamlit==1.55.0
torch==2.10.0
transformers==4.57.6
yfinance==1.2.0
